Huge Changes
Focusing on last text input now automatically adds another list item to
the end of the list.

Checkboxes are properly styled without too much hacking.

Checking items moves them to a "done" list where the text is styled with
a strikethrough. Items in the "done" list can be unchecked and moved
back to the main list.